feat(...): add backend_factory to faiss and scipy_stats extension
